refactor: move paymentcard into separate file
