+++ frontend/src/features/public-form/components/FormPaymentPage/components/ProductPaymentItemDetailsBlock.tsx
@@ -1,14 +1,6 @@
 import { useEffect, useState } from 'react'
 import { useFormContext } from 'react-hook-form'
-import {
-  Box,
-  Divider,
-  Flex,
-  FormControl,
-  Stack,
-  Text,
-  useDisclosure,
-} from '@chakra-ui/react'
+import { Divider, Flex, FormControl, Stack, Text } from '@chakra-ui/react'
 
 import { PAYMENT_PRODUCT_FIELD_ID } from '~shared/constants'
 import {
@@ -19,102 +11,9 @@ import {
 import { calculatePrice } from '~shared/utils/paymentProductPrice'
 import { centsToDollars, formatCurrency } from '~shared/utils/payments'
 
-import { BxsChevronDown } from '~assets/icons/BxsChevronDown'
-import Button from '~components/Button'
-import Checkbox from '~components/Checkbox'
 import FormErrorMessage from '~components/FormControl/FormErrorMessage'
-import Radio from '~components/Radio'
 
-import PaymentQuantityModal from './PaymentQuantityModal'
-
-const ItemQuantityButton = ({
-  product,
-  onClick,
-}: {
-  product: ProductItem
-  onClick: () => void
-}) => {
-  if (!product.data.multi_qty) {
-    return <></>
-  }
-  return (
-    <Box minWidth="6.75rem" height="2.75rem" onClick={onClick}>
-      <Button
-        rightIcon={<BxsChevronDown fontSize="1.5rem" />}
-        colorScheme="secondary"
-        aria-label="Change"
-        variant="clear"
-      >
-        Qty: {product.quantity}
-      </Button>
-    </Box>
-  )
-}
-
-const PaymentItem = ({
-  product,
-  colorTheme,
-  onItemChange,
-  isMultiSelect,
-}: {
-  product: ProductItem
-  colorTheme: FormColorTheme
-  onItemChange: (
-    productId: string,
-    isSelected: boolean,
-    selectedQuantity: number,
-  ) => void
-  isMultiSelect: boolean
-}) => {
-  const ChoiceElement = isMultiSelect ? Checkbox : Radio
-  const paymentQuantityModalDisclosure = useDisclosure()
-  return (
-    <Box w="100%">
-      <ChoiceElement
-        isChecked={product.selected}
-        onChange={() =>
-          onItemChange(product.data._id, !product.selected, product.quantity)
-        }
-        variant="fullWidth"
-        p="1rem"
-        minH="6.75rem"
-      >
-        <Box flexGrow={1}>
-          <Text textStyle="subhead-1">{product.data.name}</Text>
-          <Text textStyle="h3" mb="0.5rem">
-            {product.data.description}
-          </Text>
-          <Flex alignItems={'center'}>
-            <Box flexGrow={1} as="h2" textStyle="h3">
-              S
-              {formatCurrency(
-                Number(centsToDollars(product.data.amount_cents ?? 0)),
-              )}
-            </Box>
-            <ItemQuantityButton
-              product={product}
-              onClick={paymentQuantityModalDisclosure.onOpen}
-            />
-          </Flex>
-        </Box>
-      </ChoiceElement>
-
-      <PaymentQuantityModal
-        isOpen={paymentQuantityModalDisclosure.isOpen}
-        itemName={product.data.name}
-        onCancel={paymentQuantityModalDisclosure.onClose}
-        initialQty={product.quantity || 1}
-        onSubmit={(qty) => {
-          paymentQuantityModalDisclosure.onClose()
-          onItemChange(product.data._id, true, qty)
-        }}
-        onClose={paymentQuantityModalDisclosure.onClose}
-        minQty={product.data.min_qty}
-        maxQty={product.data.max_qty}
-      />
-    </Box>
-  )
-}
+import { PaymentCard } from './PaymentCard'
 
 const isProductSelected = (productItems: Array<ProductItem>) => {
   const isSelected = productItems.some((item) => item.selected)
@@ -198,7 +97,7 @@ export const ProductPaymentItemDetailsBlock = ({
       <FormControl isInvalid={!!errors.payment_products}>
         <Stack spacing="1rem" divider={<Divider />}>
           {productItems.map((product, idx) => (
-            <PaymentItem
+            <PaymentCard
               key={product.data._id || idx}
               product={product}
               colorTheme={colorTheme}
